requirements.txt
numpy
torch
pandas
pyyaml
yfinance
